+++ src/django_elasticsearch_dsl_drf/filter_backends/__init__.py
@@ -7,6 +7,7 @@ from .filtering import (
     FilteringFilterBackend,
     GeoSpatialFilteringFilterBackend,
     IdsFilterBackend,
+    NestedFilteringFilterBackend,
 )
 from .ordering import (
     DefaultOrderingFilterBackend,